Updated the design of the translation page

+++ listener.py
@@ -76,8 +76,9 @@ if __name__ == "__main__":
     import time
     st = time.time()
     files = glob.glob('audio_stream/clips/*')
-    for f in files:
-        os.remove(f)
+    if len(files)!=0:   
+        for f in files:
+            os.remove(f)
 
 
     for x in range (cycles):
